blog/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = []
readme = "../README.md"
license = "../LICENSE"

[tool.poetry.dependencies]
python = "^3.10"
bleach = "4.1.0"
celery = "5.1.2"
click = "7.1.2"
email-validator = "1.1.3"
Flask = "2.0.1"
Flask-Admin = "1.5.8"
Flask-Babel = "2.0.0"
Flask-Dance = "5.0.0"
Flask-Login = "0.5.0"
Flask-Mail = "0.9.1"
Flask-Migrate = "3.1.0"
Flask-Moment = "1.0.2"
Flask-SQLAlchemy = "2.5.1"
Flask-WTF = "0.15.1"
gunicorn = "20.1.0"
Markdown = "3.3.4"
Pillow = "8.3.0"
# psycopg2-binary = "2.8.6"
pyotp = "2.6.0"
PyQRCode = "1.2.1"
python-dotenv = "0.19.0"
python-slugify = "5.0.2"
redis = "3.5.3"
sentry-sdk = "1.3.1"
SQLAlchemy = "1.4.23"
SQLAlchemy-Utils = "0.37.8"
urllib3 = "1.25.11"

[tool.poetry.group.dev.dependencies]
pyproject-flake8 = "6.0.0"
mypy = "0.910"
pytest = "6.2.5"
betamax = "0.8.1"

[tool.flake8]
exclude = ["./migrations", "venv", ".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
ignore = ["E402", "W503"]
max-line-length = 100

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
